fof8-ml/pyproject.toml
[project]
name = "fof8-ml"
version = "0.1.0"
requires-python = ">=3.12,<3.13"
dependencies = [
    "polars>=1.0.0",
    "jupyter>=1.1.1",
    "altair>=5.4.1",
    "seaborn>=0.13.2",
    "hvplot>=0.12.2",
    "bokeh>=3.9.0",
    "pyarrow>=15.0.0",
    "numpy>=2.4.4",
    "scipy>=1.17.1",
    "xgboost",
    "lightgbm",
    "hydra-core",
    "mlflow>=2.0.0",
    "fof8-core",
    "scikit-learn",
    "hydra-optuna-sweeper",
    "catboost",
    "shap>=0.51.0",
    "panel>=1.8.10",
    "jupyter-bokeh>=4.0.5",
    "torch",
    "torchvision>=0.26.0",
    "optuna==4.8.0",
    "sqlalchemy<2.0.0",
    "alembic>=1.6.5",
    "dagshub",
    "dvc",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/fof8_ml"]

[tool.uv]
package = true

[[tool.uv.index]]
name = "pytorch-cu130"
url = "https://download.pytorch.org/whl/cu130"
explicit = true

[tool.uv.sources]
torch = { index = "pytorch-cu130" }
torchvision = { index = "pytorch-cu130" }
fof8-core = { workspace = true }
